ControlBoard: now uses singleton to store and retrieve device pointers.
Now each opened motionControl device is stored by the handler to be retrieved afterward.
A controlBoardWrapper can attach itself to an already opened devices.
If the devices it is looking for do not exists yet, it closes down, no magic yet.
Be sure to put the wrapper-only device at the end, after all motor control devices.