Fix bug introduced in pr #182
In pr #182, it was introduced the storing of yarp device pointers
in the singleton. Unfortunatly the key chosen was not scoped, but
was instead only the part name. This was causing failures when tryng to
run different robots that shared part names. Fixing the issue by
actually using a scoped (i.e. globally unique) key for storing
the pointers in the singleton.