Fix deletion of model in Gazebo 8
Streamlining of the disconnection event: simply resetting the
Connection shared pointer is enough to disconnect an event
since at least Gazebo 2. So the code was cleaned up from all
the call to the Disconnect* functions are all the unnecessary
check if a connection pointer is valid before resetting it.

Fix https://github.com/robotology/gazebo-yarp-plugins/issues/282 .