[controlboard] Add three-parameters variant of setRefTorques
This method was implemented in YARP (and in the ControlBoardWrapper)
in https://github.com/robotology/yarp/pull/349 , but it was never
implemented in Gazebo. Unfortunatly, as the setRefTorques result
in message sent only on the streaming port, the call to setRefTorques
was correctly returning true, and the only error was a print in the
ControlBoardWrapper running inside Gazebo, so the failure was quite silent.