Changed how joint control is managed.
Until now gazebo messages were sent to gazebo using Gazebo publish/subscribe topic.

With this change, we integrate directly the gazebo PIDs inside the code and we only command Force (or torque) to the joint.
As a second step we might think of removing the gazebo PID, but for now it seems a good compromise.

Fixes https://github.com/robotology/gazebo-yarp-plugins/issues/165