Avoid useless linking in gazebo_yarp_singleton
`${YARP_LIBRARIES}` has the peculiar behavior that the contained targets depends on the arguments passed
`find_package(YARP ...)` in this project. In particular, if (and only if) `GAZEBO_YARP_PLUGINS_HAS_OPENCV` is enabled,
then  also the `YARP::YARP_cv` component of yarp is included, that is not required at all. 

For this reason, it is better to just link the libraries we actually depend on, otherwise it will be impossible to use `gazebo_yarp_singleton` 
in a downstream project that do not have the  `YARP::YARP_cv` target defined.