Avoid to append GAZEBO_CXX_FLAGS to CMAKE_CXX_FLAGS
GAZEBO_CXX_FLAGS was added to CMAKE_CXX_FLAGS in https://github.com/robotology/gazebo-yarp-plugins/pull/190
to support compilation against Gazebo 6 that required C++11, and unfortunately since then Gazebo did not switch to a modern
way of specifying C++ standard requirement in consuming libraries. Now that YARP requires C++14, passing GAZEBO_CXX_FLAGS 
(that contains -std=c++11) basically bypass the CMake logic to compute the correct C++ standard with which to compile the plugins,
and forces to compile them with C++11. 

As the only reason for which GAZEBO_CXX_FLAGS was added to CMAKE_CXX_FLAGS was to enable C++11 support, and as that is already 
passed as a requirement by YARP, I think we can just remove it.

Related SDFormat PR: https://github.com/osrf/sdformat/pull/251 .