Add integration test for gazebo_yarp_controlboard plugin
Add basic integration test for the position control and the velocity control.
The test simulates a simple one dof model (a pendulum) with a joint actuated
by a gazebo_yarp_controlboard, and uses the remote_controlboard and the controlboardwrapper2
to specify a position first and velocity setpoint later, and verify that the
simulated robot reaches the requested setpoints. To avoid the need to
run the yarpserver externally, the test uses YARP in local mode.